CONN_2167: updates to reflect new WSDL changes and sonar violation for _10 package
