+++ Product/Production/Adapters/DocumentDataSubmission_a0/src/main/java/gov/hhs/fha/nhinc/docdatasubmission/adapter/AdapterDocDataSubmissionUnsecured.java
@@ -26,7 +26,7 @@
  */
 package gov.hhs.fha.nhinc.docdatasubmission.adapter;
 
-import gov.hhs.fha.nhinc.adapterdds.data.AdapterXDSPortType;
+import gov.hhs.fha.nhinc.adapterxds.AdapterXDSPortType;
 import gov.hhs.fha.nhinc.common.nhinccommonadapter.AdapterRegisterDocumentSetRequestType;
 import javax.annotation.Resource;
 import javax.xml.ws.BindingType;
